Refactor category management by restructuring resources
- Moved CategoryResource to a new namespace: Firefly\FilamentBlog\Resources\Categories.
- Removed deprecated getForm method from the Category model.
- Introduced new schemas for CategoryForm and CategoryInfolist to streamline form and infolist configurations.
- Added new pages for creating, editing, listing, and viewing categories.
- Implemented a PostsRelationManager for managing related posts within categories.
- Updated Blog.php to reflect the new CategoryResource path.
