+++ src/Models/Category.php
@@ -2,14 +2,10 @@
 
 namespace Firefly\FilamentBlog\Models;
 
-use Filament\Schemas\Components\Utilities\Get;
-use Filament\Schemas\Components\Utilities\Set;
-use Filament\Forms\Components\TextInput;
 use Firefly\FilamentBlog\Database\Factories\CategoryFactory;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
-use Illuminate\Support\Str;
 
 class Category extends Model
 {
@@ -27,26 +23,6 @@ class Category extends Model
     public function posts(): BelongsToMany
     {
         return $this->belongsToMany(Post::class, config('filamentblog.tables.prefix').'category_'.config('filamentblog.tables.prefix').'post');
-    }
-
-    public static function getForm()
-    {
-        return [
-            TextInput::make('name')
-                ->live(true)
-                ->afterStateUpdated(function (Get $get, Set $set, ?string $operation, ?string $old, ?string $state) {
-
-                    $set('slug', Str::slug($state));
-                })
-                ->unique(config('filamentblog.tables.prefix').'categories', 'name', null, 'id')
-                ->required()
-                ->maxLength(155),
-
-            TextInput::make('slug')
-                ->unique(config('filamentblog.tables.prefix').'categories', 'slug', null, 'id')
-                ->readOnly()
-                ->maxLength(255),
-        ];
     }
 
     protected static function newFactory()
