+++ src/Resources/Categories/Schemas/CategoryInfolist.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace Firefly\FilamentBlog\Resources\Categories\Schemas;
+
+use Filament\Infolists\Components\TextEntry;
+use Filament\Schemas\Components\Section;
+use Filament\Schemas\Schema;
+
+class CategoryInfolist
+{
+    public static function configure(Schema $schema): Schema
+    {
+        return $schema
+            ->components([
+                Section::make('Category')
+                    ->schema([
+                        TextEntry::make('name'),
+                        TextEntry::make('slug'),
+                    ])
+                    ->columns(2)
+                    ->icon('heroicon-o-square-3-stack-3d'),
+            ]);
+    }
+}
