--- src/Resources/CategoryResource.php
@@ -1,100 +1,0 @@
-<?php
-
-namespace Firefly\FilamentBlog\Resources;
-
-use Filament\Schemas\Schema;
-use Filament\Tables\Columns\TextColumn;
-use Filament\Actions\EditAction;
-use Filament\Actions\ViewAction;
-use Filament\Actions\DeleteAction;
-use Filament\Actions\BulkActionGroup;
-use Filament\Actions\DeleteBulkAction;
-use Filament\Schemas\Components\Section;
-use Firefly\FilamentBlog\Resources\CategoryResource\Pages\ListCategories;
-use Firefly\FilamentBlog\Resources\CategoryResource\Pages\EditCategory;
-use Firefly\FilamentBlog\Resources\CategoryResource\Pages\ViewCategory;
-use Filament\Infolists\Components\TextEntry;
-use Filament\Resources\Resource;
-use Filament\Tables;
-use Filament\Tables\Table;
-use Firefly\FilamentBlog\Models\Category;
-use Firefly\FilamentBlog\Resources\CategoryResource\RelationManagers\PostsRelationManager;
-
-class CategoryResource extends Resource
-{
-    protected static ?string $model = Category::class;
-
-    protected static string | \BackedEnum | null $navigationIcon = 'heroicon-o-squares-plus';
-
-    protected static string | \UnitEnum | null $navigationGroup = 'Blog';
-
-    protected static ?int $navigationSort = 1;
-
-    public static function form(Schema $schema): Schema
-    {
-        return $schema
-            ->components(Category::getForm());
-    }
-
-    public static function table(Table $table): Table
-    {
-        return $table
-            ->columns([
-                TextColumn::make('name')
-                    ->searchable(),
-                TextColumn::make('slug'),
-                TextColumn::make('posts_count')
-                    ->badge()
-                    ->counts('posts'),
-                TextColumn::make('created_at')
-                    ->dateTime()
-                    ->sortable()
-                    ->toggleable(isToggledHiddenByDefault: true),
-                TextColumn::make('updated_at')
-                    ->dateTime()
-                    ->sortable()
-                    ->toggleable(isToggledHiddenByDefault: true),
-            ])
-            ->filters([
-                //
-            ])
-            ->recordActions([
-                EditAction::make(),
-                ViewAction::make(),
-                DeleteAction::make(),
-            ])
-            ->toolbarActions([
-                BulkActionGroup::make([
-                    DeleteBulkAction::make(),
-                ]),
-            ]);
-    }
-
-    public static function infolist(Schema $schema): Schema
-    {
-        return $schema->components([
-            Section::make('Category')
-                ->schema([
-                    TextEntry::make('name'),
-                    TextEntry::make('slug'),
-                ])->columns(2)
-                ->icon('heroicon-o-square-3-stack-3d'),
-        ]);
-    }
-
-    public static function getRelations(): array
-    {
-        return [
-            PostsRelationManager::class,
-        ];
-    }
-
-    public static function getPages(): array
-    {
-        return [
-            'index' => ListCategories::route('/'),
-            'edit' => EditCategory::route('/{record}/edit'),
-            'view' => ViewCategory::route('/{record}'),
-        ];
-    }
-}
